Version 1.09.03
1. Repair Contracts totally redone.
2. You now need a few dependecies to run Mission
Controller 2s new Repair Contracts.
a. Kerbal Attachment System (NOT INCLUDED)
b. Kerbal EVA Resource Transfer (INCLUDED)
c. Community Resource Pack (INCLUDED)

New Repair Contracts.

You no longer have to DOCK with a satellite.
The 2 objectives are to Repair, and to have 1 SpareParts.

repairParts have been replaced with SpareParts.

How to do repair contracts.
1. Must have satellite in space with Repair Panel.
2. Must have a least 1 SpareParts on your Kerbal or Vessel.
3. Launch vessel to the Target Satellite.  You can dock if you wish but
its not needed.
4. GO EVA and grab the Repair Canister and place it on your back using
KAS Grab.  (or use the EVA X from Universal Storage)
5. EVA to the RepairPanel on the satellite.
6. Right click the RepairPanel and select Open Transfer GUI.
7. Transfer SpareParts to Repair Panel by selecting OUT.
8.  Turn off the Open Transfer GUI.
9. Right click Repair Panel and Select Test System option (please note
that if this is a pre 1.09 satellite you will actually have to open the
Repair Panel Doors, then take control of the satellite to see the Test
Systems button)
10.  Once the system is preped (message will tell you) Right click the
repair Panel Again and select start repairs!

Thats it if you are at the correct satellite the contract will be
completed!

You can still dock with the satellite if you wish, I had to do some
special coding to make this work, but if you dock with the vessel MCE
will change what needs to be changed under the hood for this to work
with New vessel ID's you get when you dock with something.