Stage 1 Of 1.9 Codes Update
stage one update done, getting all parameters in place to be only
achieved by new vessels is on way.  Also bit the bullet and looked at
other contract mods see how they are handling changing of vessels and
having Parameters and goals not be complete.

Even though it was not perfect I decided to use FinePrints version of
how he does it.  using OnVesselchange and VesselReady.  It's not perfect
which is why I added one more check.  (testing under
timeCountdownOrbits) and that is If vessel launch is Greater than
Contract Acceptance continue with code.

What this means is that basically a lot of the parameters (with a few
acceptions) will now require a NEW vessel for every new contract.  It
works very well, and is one of the biggest things to stop unwanted
parameters or Contracts being completed by other vessel when switching.