Added Hardcore Mode For contracts
What this does is turn on the Vessel Must Survive
Parameter for All Satellite contracts and Civilian Contracts.  If your
vessel is destroyed while in this mode the contract will fail and you
will lose a huge amount of Funds.

This works pretty well, but I think the reason its not used much is
because bascially if you have an active contract and are messing around
with another vessel and its destroyed it could set this trigger off.

If have added some checks but it can still happen.  Mostly any vessel
built before contract is accepted is safe.  Any vessel launched after IS
NOT!